Pad strides to 16 bytes.
