+++ src/util.rs
@@ -46,3 +46,29 @@ fn sanity() {
   let a: AlignedArray<_> = AlignedArray([0u8; 3]);
   assert!(a.array.as_ptr() as usize % 16 == 0);
 }
+
+pub trait Fixed {
+    fn floor_log2(&self, n: usize) -> usize;
+    fn ceil_log2(&self, n: usize) -> usize;
+    fn align_power_of_two(&self, n: usize) -> usize;
+    fn align_power_of_two_and_shift(&self, n: usize) -> usize;
+}
+
+impl Fixed for usize {
+    #[inline]
+    fn floor_log2(&self, n: usize) -> usize {
+        self & !((1 << n) - 1)
+    }
+    #[inline]
+    fn ceil_log2(&self, n: usize) -> usize {
+        (self + (1 << n) - 1).floor_log2(n)
+    }
+    #[inline]
+    fn align_power_of_two(&self, n: usize) -> usize {
+        self.ceil_log2(n)
+    }
+    #[inline]
+    fn align_power_of_two_and_shift(&self, n: usize) -> usize {
+        (self + (1 << n) - 1) >> n
+    }
+}
